[Add] Add link-prefix.js and site-meta.js for constants
