+++ src/constants/link-prefix.js
@@ -1,0 +1,8 @@
+export const LINK_PREFIX = {
+  ARTICLE: '/a/',
+  INTERACTIVE_ARTICLE: '/i/',
+  CATEGORY: '/category/',
+  TAG: '/tag/',
+  TOPIC: '/topic/',
+  TOPICS: '/topics/'
+}
