+++ src/constants/site-meta.js
@@ -1,0 +1,8 @@
+export const SITE_META = {
+  DESC: '《報導者》是由「財團法人報導者文化基金會」成立的非營利網路媒體，致力於公共領域的深度報導及調查報導，為讀者持續追蹤各項重要議題。我們秉持開放參與的精神，結合各種進步價值與公民力量，共同打造多元進步的社會與媒體環境。',
+  URL: 'https://www.twreporter.org/',
+  URL_NO_SLASH: 'https://www.twreporter.org',
+  LOGO: 'https://www.twreporter.org/asset/logo.png',
+  KEYWORDS: '報導者,twreporter,新聞,調查報道,原創報道,深度報道,環境,議題,評論,媒體,新媒體,台灣,兩岸,文化'
+}
+
